+++ src/Schema/Directives/FirstDirective.php
@@ -50,7 +50,6 @@ SDL;
      */
     public function resolveField(FieldValue $fieldValue): FieldValue
     {
-        /** @var \Illuminate\Database\Eloquent\Model $model */
         $model = $this->getModelClass();
 
         return $fieldValue->setResolver(