+++ tests/Integration/Execution/MutationExecutor/BelongsToManyTest.php
@@ -59,14 +59,7 @@ class BelongsToManyTest extends DBTestCase
         id: ID!
         name: String
     }
-    ';
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->schema .= $this->placeholderQuery();
-    }
+    ' . self::PLACEHOLDER_QUERY;
 
     public function testCanCreateWithNewBelongsToMany(): void
     {