+++ tests/Unit/Schema/AST/ASTBuilderTest.php
@@ -35,7 +35,7 @@ class ASTBuilderTest extends TestCase
             baz: Boolean
         }
         ';
-        $documentAST = $this->astBuilder->build();
+        $documentAST = $this->astBuilder->documentAST();
 
         $this->assertCount(
             3,
@@ -56,6 +56,6 @@ class ASTBuilderTest extends TestCase
         ';
 
         $this->expectException(DefinitionException::class);
-        $this->astBuilder->build();
+        $this->astBuilder->documentAST();
     }
 }