+++ tests/Unit/Schema/ClientDirectiveTest.php
@@ -9,8 +9,6 @@ class ClientDirectiveTest extends TestCase
 {
     public function testReturnsDefaultDirectivesInIntrospection(): void
     {
-        $this->schema = $this->placeholderQuery();
-
         $this->assertNotNull(
             $this->introspectDirective(Directive::SKIP_NAME)
         );
@@ -21,13 +19,13 @@ class ClientDirectiveTest extends TestCase
 
     public function testCanDefineACustomClientDirective(): void
     {
-        $this->schema = '
+        $this->schema .= '
         "foo"
         directive @bar(
             "foobar"
             baz: String = "barbaz"
         ) on FIELD
-        '.$this->placeholderQuery();
+        ';
 
         $bar = $this->introspectDirective('bar');