+++ tests/Unit/Schema/Directives/CanDirectiveTest.php
@@ -156,43 +156,6 @@ class CanDirectiveTest extends TestCase
         ')->assertErrorCategory(AuthorizationException::CATEGORY);
     }
 
-    public function testWorksWithPaginateDirection(): void
-    {
-        $user = new User;
-        $user->name = UserPolicy::ADMIN;
-        $this->be($user);
-
-        $this->schema = '
-        type Query {
-            users: [User!]!
-                @can(ability: "adminOnly")
-                @paginate
-        }
-        
-        type User {
-            name: String
-        }
-        ';
-
-        $this->graphQL('
-        {
-            users(first: 1) {
-                data {
-                    name
-                }
-            }
-        }
-        ')->assertJson([
-            'data' => [
-                'users' => [
-                    'data' => [
-                        'name' => 'foo',
-                    ],
-                ],
-            ],
-        ]);
-    }
-
     public function resolveUser(): User
     {
         $user = new User;